convert it into standalone plugin
