+++ build_logic/build_info_plugin/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+  `kotlin-dsl`
+}
+
+gradlePlugin {
+  plugins {
+    create("buildInfoPlugin") {
+      id = "buildInfoPlugin"
+      implementationClass = "BuildInfoPlugin"
+    }
+  }
+}
