+++ build_logic/settings.gradle.kts
@@ -1,0 +1,14 @@
+dependencyResolutionManagement {
+  repositories {
+    google()
+    mavenCentral()
+  }
+  versionCatalogs {
+    create("libs") {
+      from(files("../gradle/libs.versions.toml"))
+    }
+  }
+}
+
+rootProject.name = "build_logic"
+include(":build_info_plugin")
