+++ gradle/libs.versions.toml
@@ -47,6 +47,7 @@ room-runtime = { group = "androidx.room", name = "room-runtime", version.ref = "
 android-application = { id = "com.android.application", version.ref = "agp" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
+build-info = { id = "buildInfoPlugin" }
 room = { id = "androidx.room", version.ref = "room" }
 
 [bundles]
