+++ settings.gradle.kts
@@ -1,4 +1,5 @@
 pluginManagement {
+  includeBuild("build_logic")
   repositories {
     google {
       content {
